Change Partitioner tests to mock at the function level
This helps out with some problems the tests previously had, which was
that they were sharing state by having the entire class contained in a
single mock context.

By switching to per-function contexts the moto mock state is reset with
each function. While that brings some new issues (mainly in making the
tests individually more verbose, because setup is duplicated in each of
them), the advantages to not sharing state are much better.

In addition to switching to the per-function mock contexts, the helper
functions from both partitioner_test and database_cleaner_test have been
merged into a single helper class, shared between them.

Minor changes were made to the database cleaner tests to conform with
the updated helper.