Rework the partition creation flow
As part of working to make this library more composable for the CLI and
for testing, the create_new_partitions logic was pulled out of the
Partitioner class, and the create_partitions function was publicized.

Related changes were made to the tests around creating partitions and to
the athena_partitioner.py

In addition to those big changes, the athena_partitioner.py logic has
been moved into a glutil.serverless_function file, to keep the logic as
part of the library.