Add limit_days option to create_partitions
We've been having a number of our create_partition lambda runs time out
because our data sets are quite large, and it takes a lot of time
to crawl S3 for all potential partitions.

The new limit_days argument allows you to limit the partitioner to only
look for new partitions that match the pattern for the last N days.

By making this switch, Partitioner.partitions_in_s3 has gone from 41
seconds with no limit to 1 second with a seven day limit on one of our
larger data sets.