Add Partitioner.partitions_to_create, for determining what partitions to create
The previous method of determining which partitions should be created
was by querying all partitions in the table, and comparing them against
a set of partitions that should be created.

This change switches to using the batch-get-partition AWS API to instead
query AWS for all partitions we want to create, instead of all
partitions period, and comparing that against the set of partitions to
create.

This works because the batch-get-partition api returns only existing
partitions, and ignores any partitions that don't exist in the input.

Like the last change to add limit_days, this is an optimization to speed
up the glue crawler alternative lambda, in very limited testing against
one data set with one new partition the time for determining which
partitions to create went from ~20s to ~0.5s.